+++ packages/kit/test/apps/basics/src/routes/navigation-lifecycle/before-navigate/prevent-navigation/+page.svelte
@@ -20,5 +20,6 @@
 <a href="/navigation-lifecycle/before-navigate/prevent-navigation?x=1">self</a>
 <a href="https://google.com" target="_blank" rel="noreferrer">_blank</a>
 <a href="https://google.de">external</a>
+<!-- svelte-ignore a11y-invalid-attribute -->
 <a download href="">external</a>
 <pre>{times_triggered} {unload} {navigation_type}</pre>